BUG: fixed bug, changed EffectSoilFertilitySalinityStress
Remove attribute of Simul.pas EffectSoilFertilitySalinityStress and
changed to Getters/Setters of the Simulation%EffectStress attribute

Now there is a bug for some other pixels (~10 in total with irrigation).
The deviation in output is always at the same place (index 68, line 97)
and deviates from 0.1 (e.g. 9.3->9.4). So new testing works (but fails
with stricter tolerance)