TST: only allow 1 small deviation in 9 cases (otherwise 0)
This modifies the Europe test so that a single off-by-one error
in the least significant digit is allowed in 9 specific cases.
For all other cases the strict criterion remains in place.

The single deviations in these 9 cases are not introduced in this
branch, but were already present for a longer time, presumably
due to this commit (as identified by Louise):
fda50076f0bb77eeee6e39d6da6a8afb056e6c20