BUG: multiple string pointers need multiple 'saved' objects
Before this, the ClimateRecord.FromString and ClimateRecord.ToString
at the end of the CompleteClimateDescription procedure in
interface_global.pas would always contain the same string, even though
they are different on the Fortran side. The reason is that there is only
one 'saved' f_string object, which gets overwritten by the second
string2pointer call. The solution is to have a twostrings2twopointers
function which 'saves' a f_string1 and a f_string2 object, one for each
argument.