ENH: add possibility for Pascal lib + Fortran exe
With the (default) FORTRAN_EXE=0 make option, the build process
is as before (first the Fortran library is built, and then the
Pascal binary).

Now, by passing FORTRAN_EXE=1 to make, the Pascal code is compiled
into a "libPlugInBare.so" library instead. And then a little Fortran
program is compiled, which links to this library (which, in turn,
is linked to the Fortran-based "libaquacrop.so" library).