BUG: only use CCiActual globvar in DetermineCCi(GDD)
In the calls to DetermineCCi & DetermineCCiGDD, the CCiActual
global variable is passed as a "VAR" argument. These routines
are then modifying this variable, which is OK, except that it
seems that some routine(s) called by DetermineCCi(GDD) rely
on the updated value of this global variable. One way to solve
this is to not pass CCiActual as an argument, and let the
DetermineCCi(GDD) modify the global variable through the usual
getter and setter.