REV: revert to original WriteAssessmentSimulation wrappers
The omission of the first string (pchar+strlen) argument was
needed when WriteAssessmentSimulation was still being called
from Pascal. Now that this is no longer the case, it is more
consistent to apply the usual wrapping (but with a warning
in case empty strings are passed, e.g. for StrNr, which was
the problematic case).