ENH: store Temperature data and use in daily cases
The TemperatureFilefull file was being opened and read many times over
(around 13 000 times in the Perennial test, making up for the majority
of the time spent in this calculation). Storing its contents in arrays
and using them in the SumCalendarDays and GrowingDegreeDays routines
in case of a 'daily' TemperatureRecord_Datatype already provides major
speedups (from 26-30 seconds down to 12-13 seconds).