TST: use $(AQUACROP_)TMPDIR and add suffix
Now $TMPDIR is used by default as the parent location for the
(temporary) working directories. This can be changed by setting
the AQUACROP_TMPDIR environment variable. A suffix also gets
appended now, so that the directory names look like this:
"tmp<random>_aquacrop_<username>" with <random> the usual
random string of 8 characters.