BUG: avoid int8 underflow wraparound
The rounding can occasionally produce negative values (e.g. -199) which are
below the lower int8 range of -128 and which then get wrapped around to a
positive number (e.g. 57). For such negative values, however, the intended
value is 0. In some test cases (e.g. Europe_True-50-12) this leads to
small changes in the final output, so a new set of reference values will
need to be regenerated.