ENH: use ProjectInput array with whole input contents
In this way, the ProjectInput variable can be initialized
at the very beginning (before starting specific runs).
This is needed because e.g. InitializeProject and CheckForKeepSWC
already need information from the whole project file.
In this commit, also the GetNumberSimulationRuns is moved
to a more appropriate location (i.e. ac_project_input) and
the interfaces are modified accordingly.