BUG: status='replace' needed for overwriting existing files
With status='new', the resulting file handle would point to the
stderr output instead. And the status='new' option is what we should
rather be using in the 'append' case if the file does not exist.
The bug only occurred when running the AquaCrop executable more
than once without removing the resulting output files in between.